+++ lib/rito/endpoints/lol/clash_v1.rb
@@ -6,8 +6,8 @@ module Rito
       class ClashV1 < Base
         ROUTING = :platform
 
-        def players_by_summoner_id(summoner_id, region: nil)
-          get("/lol/clash/v1/players/by-summoner/#{escape(summoner_id)}", region)
+        def players_by_puuid(puuid, region: nil)
+          get("/lol/clash/v1/players/by-puuid/#{escape(puuid)}", region)
         end
 
         def team(team_id, region: nil)
@@ -16,6 +16,10 @@ module Rito
 
         def tournaments(region: nil)
           get('/lol/clash/v1/tournaments', region)
+        end
+
+        def tournament(tournament_id, region: nil)
+          get("/lol/clash/v1/tournaments/#{escape(tournament_id)}", region)
         end
 
         def tournament_by_team(team_id, region: nil)