+++ lib/rito/endpoints/valorant.rb
@@ -52,6 +52,17 @@ module Rito
         end
       end
 
+      class ConsoleRankedV1 < Base
+        ROUTING = :valorant_console_platform
+
+        def leaderboard(act_id, platform_type:, size: nil, start_index: nil, region: nil)
+          params = { 'platformType' => platform_type }
+          params['size'] = size if size
+          params['startIndex'] = start_index if start_index
+          get("/val/console/ranked/v1/leaderboards/by-act/#{escape(act_id)}", region, params: params)
+        end
+      end
+
       class StatusV1 < Base
         ROUTING = :valorant_platform
 
@@ -78,6 +89,10 @@ module Rito
 
       def console_matches
         Endpoints::Valorant::ConsoleMatchV1.new(@client)
+      end
+
+      def console_ranked
+        Endpoints::Valorant::ConsoleRankedV1.new(@client)
       end
 
       def ranked